% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icanhazip.r
\name{icanhazip}
\alias{icanhazip}
\title{Use icanhazip to retrieve public IP address}
\usage{
icanhazip()
}
\description{
This service can return both IPv4 and IPv6 addresses
and also operates over \code{https}.
}
