% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myip-package.r
\docType{package}
\name{myip-package}
\alias{myip-package}
\title{Tools to Determine Your Public 'IP' Address}
\description{
It is often necessary to know the 'IPv4' and/or 'IPv6'
external addresses of a workstation for 'API' calls. Functions are
provided to use one or more external services to accomplish this task
along with a wrapper function that can return results from multiple
services or use a random service.
}
\author{
Bob Rudis (@hrbrmstr)
}

